extract common functionality for timeline resolvers
